+++ src/main.js
@@ -55,6 +55,7 @@ Vue.component('font-awesome-icon', FontAwesomeIcon);
             })
         })
     })();
+    api.head(store.state.config.api);
     new Vue({
         store,
         i18n,