+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Change Log
 
+### 2019-07-24
+
+1. Add copy back
+2. Merge daovoice into `usr.js`
+3. Add copy btn
+4. Add markdown switch
+5. Update success page's clipboard include
+
 ### 2019-07-23
 
 1. Remove prismjs
 2. Add highlightjs
 3. Add line numbers to highlightjs
 4. Bug fix of duplicate syntax when language change
+5. Update baseUrl, remove `protocol` & `base_url` from config.json
+6. Update highlight's register
 
 ### 2019-07-22